restructure the files to avoid circular imports and make it cleaner
