+++ petpals/static/style.css
