+++ requirements.txt
@@ -6,7 +6,9 @@ colorama==0.4.4
 dnspython==2.2.1
 email-validator==1.1.3
 Flask==2.0.3
+Flask-SQLAlchemy==2.5.1
 Flask-WTF==1.0.0
+greenlet==1.1.2
 idna==3.3
 iniconfig==1.1.1
 itsdangerous==2.1.0
@@ -18,6 +20,7 @@ py==1.11.0
 pycodestyle==2.8.0
 pyparsing==3.0.7
 pytest==7.0.1
+SQLAlchemy==1.4.34
 toml==0.10.2
 tomli==2.0.1
 Werkzeug==2.0.3
